Change int/long variables to [u_]int{16,32,64}_u

+++ solutions/project-euler/001/001.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
-int main() {
-    int i, sum = 0;
+int16_t main() {
+    int16_t i, sum = 0;
     for (i = 0; i < 1000; i++)
         if (i % 3 == 0 || i % 5 == 0)
             sum += i;
