+++ solutions/project-euler/004/004.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <string.h>
 
-int capicua(int n) {
-    int i;
+int16_t capicua(int16_t n) {
+    int16_t i;
     char s[6];
     snprintf(s, sizeof(6), "%d", n);
     for (i = 0; i < strlen(s) / 2; i++)
@@ -11,8 +11,8 @@ int capicua(int n) {
     return 1;
 }
 
-int main() {
-    int i, j, g = 1;
+int16_t main() {
+    int16_t i, j, g = 1;
     for (i = 100; i <= 999; i++)
         for (j = 100; j <= 999; j++)
             if (i * j > g && capicua(i * j))
