+++ solutions/spoj-br/contrato/contrato.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
-int main() {
-    int n, s, t;
+int16_t main() {
+    int16_t n, s, t;
     while ((t = getchar()) != '0') {
         getchar();
         s = 0;
 
         while ((n = getchar()) != '\n') {
             if (n != t) {
-                s += n - '0';  // converts char to int and then sum to s
+                s += n - '0';  // converts char to int16_t and then sum to s
                 if (s != 0) {
                     printf("%c", n);
                 }
