+++ solutions/uri/1020/1020.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
-int main() {
-    int t, a = 0, m = 0, d = 0;
+int16_t main() {
+    int16_t t, a = 0, m = 0, d = 0;
 
     while (scanf("%d", &t) != EOF) {
         if (t >= 365) {
