+++ solutions/uri/1060/1060.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
-int main() {
+int16_t main() {
     float n;
-    int c = 0;
+    int16_t c = 0;
     while (scanf("%f", &n) != EOF) {
         if (n > 0) {
             c++;
