+++ solutions/uri/1151/1151.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
-int main() {
-    int n, a, b, x, i;
+int16_t main() {
+    int16_t n, a, b, x, i;
 
     scanf("%d", &n);
